# backend/src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Travel Manager API",
        "description": "Gerenciamento de viagens corporativas",
        "version": "1.0.0"
    },
    "paths": {
...
        "/travel-orders": {
            "get": {
                "tags": [
                    "Travel Orders"
                ],
                "summary": "Get list of travel orders",
                "operationId": "5afa5082a060eebf9b85e29ea5ea0c6a",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by status",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "destino",
                        "in": "query",
                        "description": "Filter by destination",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "data_ida",
                        "in": "query",
                        "description": "Filter by start date",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "data_volta",
                        "in": "query",
                        "description": "Filter by end date",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/TravelOrder"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Travel Orders"
                ],
                "summary": "Create a new travel order",
                "operationId": "3c6f3eac075e217221019dcd1945c0d3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "destino",
                                    "data_ida",
                                    "data_volta"
                                ],
                                "properties": {
                                    "destino": {
                                        "type": "string",
                                        "example": "New York"
                                    },
                                    "data_ida": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-10-01"
                                    },
                                    "data_volta": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-10-10"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Travel order created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TravelOrder"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/travel-orders/{id}": {
            "get": {
                "tags": [
                    "Travel Orders"
                ],
                "summary": "Get a specific travel order by ID",
                "operationId": "71a313f58ac32d47e8ba27afd9b99adb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the travel order",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TravelOrder"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Travel order not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/travel-orders/{id}/status": {
            "patch": {
                "tags": [
                    "Travel Orders"
                ],
                "summary": "Update the status of a travel order",
                "operationId": "f023022f342c5cb79ee937cece017c25",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the travel order",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "aprovado",
                                            "cancelado"
                                        ],
                                        "example": "aprovado"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TravelOrder"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden - Cannot update your own order"
                    },
                    "404": {
                        "description": "Travel order not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/travel-orders/{id}/cancel": {
            "patch": {
                "tags": [
                    "Travel Orders"
                ],
                "summary": "Cancel a travel order",
                "operationId": "4472262148acec706b7c1019cc8be07d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the travel order",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Travel order canceled successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TravelOrder"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request - Only approved orders can be canceled"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Travel order not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/travel-orders/notifications": {
            "get": {
                "tags": [
                    "Travel Orders"
                ],
                "summary": "Get user's unread notifications",
                "operationId": "dfe6e7558b03899b7b24f59cf7e13f5a",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#Notification"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "TravelOrder": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "status": {
                        "type": "string",
                        "example": "solicitado"
                    },
                    "destino": {
                        "type": "string",
                        "example": "New York"
                    },
                    "data_ida": {
                        "type": "string",
                        "format": "date",
                        "example": "2023-10-01"
                    },
                    "data_volta": {
                        "type": "string",
                        "format": "date",
                        "example": "2023-10-10"
                    },
                    "user": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 1
                            },
                            "name": {
                                "type": "string",
                                "example": "John Doe"
                            },
                            "email": {
                                "type": "string",
                                "example": "john.doe@example.com"
                            }
                        },
                        "type": "object"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-10-01T12:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-10-01T12:00:00Z"
                    }
                },
                "type": "object"
            }
        }
    },
...
}